% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.R
\name{deg.freedom2S}
\alias{deg.freedom2S}
\title{This function calculates the degrees of freedom of a two sample t test.}
\usage{
deg.freedom2S(n1, n2, s1, s2, equal.variance)
}
\arguments{
\item{n1}{: sample size of the first group}

\item{n2}{: sample size of the second group}

\item{s1}{: standard deviation of the first group}

\item{s2}{: standard deviation of the second group}

\item{equal.variance}{: TRUE if qual variance assumption is made. This should be based on the results of var.test(x,y).}
}
\value{
degrees of freedom
}
\description{
This function calculates the degrees of freedom of a two sample t test.
}
\examples{
# Suppose we have the yield of corn for two types of fertilizer, A and B.
# Test to see whether there is a difference in the means of yield produced, from application of both types of fertilizers.
# H0: UA = UB , H1: UA != UB

yieldA <- c(452, 874, 554, 447, 356, 754 , 558, 574, 664, 682, 547, 435, 245)
yieldB <- c(546, 547, 774, 465, 459, 665, 467, 365, 589, 534, 456, 651, 654, 665, 546, 537)

# Perform an F test to compare the variances of the two samples
var.test(yieldA, yieldB) ## Results of this test indicate that we can assume equal variance

# Run the deg.freedom2S function to calculate the degrees of freedom
n1 <- length(yieldA)
n2 <- length(yieldB)
s1 <- sd(yieldA)
s2 <- sd(yieldB)
equal.variance <- TRUE

deg.freedom2S(n1, n2, s1, s2, equal.variance)
}
