% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGeo2xl.R
\name{r2xl}
\alias{r2xl}
\title{This function is used to to generate summary tables and export them into excel, in a format that the client can easily understand and work with.}
\usage{
r2xl(data, var, xlab, gvars, grp_titles, wbname, sr1, sr2, qt)
}
\arguments{
\item{data}{The data to be used}

\item{var}{The variable that holds the question}

\item{xlab}{The question labels for the main variables}

\item{gvars}{The variables to be used for crosstabs}

\item{grp_titles}{The question labels for the crosstab variables}

\item{wbname}{Name of the workbook where the results will be saved}

\item{sr1}{The row that will contain the question}

\item{sr2}{The row where the table will start}

\item{qt}{This indicates the question type}
}
\value{
An excel workbook saved in the path of your computer
}
\description{
This function is used to to generate summary tables and export them into excel, in a format that the client can easily understand and work with.
}
