% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGrafu.R
\name{rGraf}
\alias{rGraf}
\title{This function is used to to generate summary tables and graphs showing frequencies and percentages.}
\usage{
rGraf(data, group_var, xlab, flip, col, theme, title)
}
\arguments{
\item{data}{The data to be used}

\item{group_var}{The variable that you want to analyse}

\item{xlab}{The x axis label}

\item{flip}{TRUE/FALSE: Whether the graph should be flipped or not}

\item{col}{Colour to be used in the graph}

\item{theme}{The theme you want to use for your graphs}

\item{title}{The title of your graph}
}
\value{
A table and a graph
}
\description{
This function is used to to generate summary tables and graphs showing frequencies and percentages.
}
\examples{
## Specify the theme

theme <- theme(legend.position = "right",
         legend.direction = "vertical",
         legend.title = element_blank(),
         plot.title = element_text( size = rel(1.6), hjust = 0.5),
         plot.subtitle = element_text(size = rel(1.4), hjust = 0.5),
         axis.text.x = element_text(size =rel(1.4),angle = 0),
         axis.text.y = element_text(size =rel(1.4),angle = 0),
         axis.title = element_text( size = rel(1.5)),
         panel.background = element_rect(fill = NA))

## Call the function

rGraf(data = wafanyikazi,group_var = Gender, xlab = "Gender",flip = FALSE,col= "#154360", theme = theme, title = "Distribution of Gender")

}
