% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGrafu.R
\name{rGraf}
\alias{rGraf}
\title{This function is used to to generate summary tables and graphs showing frequencies and percentages.}
\usage{
rGraf(data, group_var, xlab, flip, theme, title)
}
\arguments{
\item{data}{The data to be used}

\item{group_var}{The variable that you want to analyse}

\item{xlab}{The x axis label}

\item{flip}{TRUE/FALSE: Whether the graph should be flipped or not}

\item{theme}{The theme you want to use for your graphs}

\item{title}{The title of your graph}
}
\value{
A table and a graph
}
\description{
This function is used to to generate summary tables and graphs showing frequencies and percentages.
}
\examples{
rGraf(data = wafanyikazi,group_var = Gender, xlab = "Gender",flip = FALSE, theme = theme_bw(), title = "Distribution of Gender")

}
